refactor

+++ Tryitter/Repository/PostRepository.cs
@@ -1,4 +1,5 @@
 using Tryitter.Model;
+using Tryitter.Views;
 
 namespace Tryitter.Repository;
 
@@ -12,52 +13,50 @@ public class PostRepository : IPostRepository
     _context = context;
   }
 
-  public IEnumerable<PostModel> GetAll()
+  public IEnumerable<PostView> GetAll()
   {
-    return _context.posts.Join(
-      _context.users,
-      post => post.User.UserId,
-      user => user.UserId,
-      (post, user) => new PostModel()
+    return (
+      from post in _context.posts
+      join module in _context.modulos on post.User.Modulo equals module
+      orderby post.CreatedAt descending
+      select new PostView()
       {
-        User = user,
+        PostId = post.PostId,
         Content = post.Content,
-        CreatedAt = post.CreatedAt,
         Likes = _context.likes.Where(x => x.PostId == post.PostId).Count(),
-        PostId = post.PostId
+        User = PostView.ConvertUserModel(post.User, module)
       }
-    ).OrderByDescending(x => x.CreatedAt);
+    );
   }
   public PostModel? GetById(int id)
   {
     return _context.posts.Find(id);
   }
-  public IEnumerable<PostModel>? GetAllByEmail(string email)
-  {
-    return GetAll().Where(posts => posts.User.Email == email); /* Isso aqui é um array */
-    // corrigir o find da L25
-  }
+  // public IEnumerable<PostModel>? GetAllByEmail(string email)
+  // {
+  //   return GetAll().Where(posts => posts.User.Email == email); /* Isso aqui é um array */
+  //   // corrigir o find da L25
+  // }
   public void Create(int userId, PostModel newPost)
-  // Verificar se não está faltando algum check
   {
     var user = _context.users.Find(userId);
-    // System.Console.WriteLine(user.Posts.Count());
     newPost.User = user;
     _context.posts.Add(newPost);
     _context.SaveChanges();
   }
-  public void Update(int id, string newPostBody)
-  // Verificar se são satisfeitas e/ou necessárias condições e validações
-  {
-    var post = GetById(id);
-    if (post == null) throw new ArgumentNullException($"PostId = {id}", "Not found");
+  // public void Update(int id, string newPostBody)
+  // // Verificar se são satisfeitas e/ou necessárias condições e validações
+  // {
+  //   var post = GetById(id);
+  //   if (post == null) throw new ArgumentNullException($"PostId = {id}", "Not found");
 
-    // _context.Entry(post).CurrentValues.SetValues(newPostBody);
-    _context.SaveChanges();
-  }
-  public void Delete(int id)
+  //   // _context.Entry(post).CurrentValues.SetValues(newPostBody);
+  //   _context.SaveChanges();
+  // }
+  public void Delete(int id, int userId)
   {
     var post = GetById(id);
+    if (post.UserId != userId) throw new InvalidOperationException("Invalid");
     if (post == null) throw new ArgumentNullException($"PostId = {id}", "Not found");
 
     _context.posts.Remove(post);
